stem-assessment-generator/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pypdf2==3.0.1
langchain==0.1.6
langchain-openai==0.0.5
chromadb==0.4.22
openai==1.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
tiktoken==0.5.2
jinja2==3.1.3
httpx==0.27.0
reportlab==4.0.9
python-docx==1.1.0
markdown2==2.4.10
pdfplumber==0.10.3
pymupdf==1.23.8
textstat==0.7.3
scikit-learn>=1.3.2
numpy>=1.25.0

# Optional optimization dependencies
redis==5.0.1  # For distributed caching (optional)
asyncio-throttle==1.0.2  # For advanced rate limiting (optional)
psutil==5.9.6  # For memory monitoring (optional)
